+++ src/TemporaryAccessService.php
@@ -3,11 +3,11 @@
 namespace Erdemkeren\TemporaryAccess;
 
 use Carbon\Carbon;
-use Erdemkeren\TemporaryAccess\Contracts\TokenInterface;
+use Erdemkeren\TemporaryAccess\Token\TokenInterface;
 use Erdemkeren\TemporaryAccess\Contracts\AccessTokenInterface;
-use Erdemkeren\TemporaryAccess\Contracts\TokenGeneratorInterface;
 use Erdemkeren\TemporaryAccess\Contracts\AccessTokenRepositoryInterface;
 use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableContract;
+use Erdemkeren\TemporaryAccess\Token\TokenGenerator\TokenGeneratorInterface;
 
 final class TemporaryAccessService
 {